+++ src/alphatamp/approaches/feasibility_classifiers/oracle_feasibility_classifier.py
@@ -1,0 +1,93 @@
+"""A classifier that uses oracle knowledge to classify abstract plans."""
+
+from typing import Hashable, TypeVar
+
+from bilevel_planning.structs import (
+    RelationalAbstractState,
+    SesameModels,
+)
+from prbench.envs.geom2d.object_types import CRVRobotType, RectangleType
+from relational_structs.object_centric_state import ObjectCentricState
+from relational_structs.pddl import (
+    GroundAtom,
+    Predicate,
+)
+
+_S = TypeVar("_S", bound=Hashable)  # abstract state
+_A = TypeVar("_A", bound=Hashable)  # abstract action
+
+
+class OracleAbstractPlanClassifier:
+    """A classifier that uses oracle knowledge to classify abstract plans."""
+
+    def __init__(
+        self,
+        env_models: SesameModels,
+    ) -> None:
+        # super().__init__(abstract_successor_function, seed)
+        self._env_models = env_models
+
+    def validate_plan(
+        self,
+        x0: ObjectCentricState | object,
+        s_plan: list[RelationalAbstractState] | list[_S],
+        _: list[_A],
+    ) -> bool:
+        """Classify abstract plans."""
+
+        # Predicates.
+        HoldingTgt = Predicate("HoldingTgt", [CRVRobotType, RectangleType])
+        HandEmpty = Predicate("HandEmpty", [CRVRobotType])
+
+        # Objects.
+        type_name_to_type = {t.name: t for t in self._env_models.types}
+        block_type = type_name_to_type["target_block"]
+        obstruction_type = type_name_to_type["rectangle"]
+        robot_type = type_name_to_type["crv_robot"]
+
+        assert isinstance(x0, ObjectCentricState)
+        (target_block,) = x0.get_objects(block_type)
+        (robot,) = x0.get_objects(robot_type)
+        obstructions = x0.get_objects(obstruction_type)
+
+        # remove target block from obstructions
+        filtered_obstructions = set()
+        for obstruction in obstructions:
+            if obstruction.name != "target_block":
+                filtered_obstructions.add(obstruction)
+
+        # Oracle abstract plan
+        empty_abstract_state_atoms: set[GroundAtom] = set()
+        empty_abstract_state_atoms.add(GroundAtom(HandEmpty, [robot]))
+        empty_abstract_state_objects = {robot, target_block} | filtered_obstructions
+        empty_abstract_state = RelationalAbstractState(
+            atoms=empty_abstract_state_atoms, objects=empty_abstract_state_objects
+        )
+
+        holding_abstract_state_atoms = set()
+        holding_abstract_state_atoms.add(GroundAtom(HoldingTgt, [robot, target_block]))
+        holding_abstract_state_objects = {robot, target_block} | filtered_obstructions
+        holding_abstract_state = RelationalAbstractState(
+            atoms=holding_abstract_state_atoms, objects=holding_abstract_state_objects
+        )
+
+        oracle_abstract_state = [empty_abstract_state, holding_abstract_state]
+
+        oracle_abstract_state_ptr = 0
+        # Classify plan only looking at state for now
+        for plan_abstract_state in s_plan:
+            assert isinstance(plan_abstract_state, RelationalAbstractState)
+            plan_atoms, plan_objects = (
+                plan_abstract_state.atoms,
+                plan_abstract_state.objects,
+            )
+            oracle_atoms, oracle_objects = (
+                oracle_abstract_state[oracle_abstract_state_ptr].atoms,
+                oracle_abstract_state[oracle_abstract_state_ptr].objects,
+            )
+
+            if plan_atoms != oracle_atoms or plan_objects != oracle_objects:
+                return False
+            oracle_abstract_state_ptr += 1
+
+        return True
